+++ src/senaite/core/catalog/indexer/contact.py
@@ -20,9 +20,9 @@
 
 from bika.lims import api
 from bika.lims.interfaces import IClient
-from bika.lims.interfaces import IContact
 from bika.lims.interfaces import ILabContact
 from bika.lims.interfaces import ISupplierContact
+from senaite.core.interfaces import IContact
 from plone.indexer import indexer
 from senaite.core.interfaces.catalog import IContactCatalog